Allowing preparedPaths to be true even if there are errors.
This change means that if there are TONS of errors, the project won't be
reprocessed over and over and over again. A simple way to see where
this would be a problem is if you had many projects that included
each other in their buildpaths. Now remove one of the foundational
projects from your workspace -- actually delete the files, not
just delete the eclipse project. the Project class will get hung up
because it continues to process all Projects because none of them
are ever "Ok".